fix: avoid rounding island spawn point to integer from double

+++ spigot/src/main/java/io/tofpu/speedbridge2/island/persistence/IslandEntity.java
@@ -5,7 +5,7 @@ public record IslandEntity(
         LocationEntity location
 ) {
     public record LocationEntity(
-            int x, int y, int z,
+            double x, double y, double z,
             float yaw, float pitch
     ) {
     }
